ft_print_char ft_print_str done

+++ Common_Core/ft_printf/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: erazumov <erazumov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/09 18:10:42 by erazumov          #+#    #+#             */
-/*   Updated: 2024/11/24 17:25:07 by erazumov         ###   ########.fr       */
+/*   Updated: 2024/11/28 13:54:55 by erazumov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ void	ft_bzero(void *dest, size_t len);
 void	*ft_memcpy(void *dest, void *src, size_t n);
 void	*ft_memmove(void *dest, void *src, size_t len);
 size_t	ft_strlcpy(char *dest, char *src, size_t size);
-int		ft_strlcat(char *dest, char *src, int size);
+int		ft_strlcat(char *dest, char *src, size_t size);
 int		ft_toupper(int c);
 int		ft_tolower(int c);
 char	*ft_strchr(const char *s, int c);
